refactor: landing components
